refactor(types): refactored flow types and moved to the appropriate dir
