+++ src/actions/sso/credentialRequest.ts
@@ -3,7 +3,7 @@ import { routeList } from 'src/routeList'
 import { JSONWebToken } from 'jolocom-lib/js/interactionTokens/JSONWebToken'
 import { CredentialRequest } from 'jolocom-lib/js/interactionTokens/credentialRequest'
 import { ThunkAction } from '../../store'
-import { CredentialVerificationSummary } from './types'
+import { CredentialVerificationSummary } from 'src/lib/interactionManager/types'
 import { InteractionChannel } from 'src/lib/interactionManager/types'
 import { Interaction } from 'src/lib/interactionManager/interaction'
 import { cancelSSO } from './'
