+++ src/ui/sso/components/checkboxCredential.tsx
@@ -2,11 +2,11 @@ import React from 'react'
 import { StyleSheet, View, Text } from 'react-native'
 import { Spacing, Typography, Colors } from 'src/styles'
 import Icon from 'react-native-vector-icons/MaterialCommunityIcons'
+import { IdentitySummary } from 'src/actions/sso/types'
 import {
   CredentialTypeSummary,
   CredentialVerificationSummary,
-  IdentitySummary,
-} from 'src/actions/sso/types'
+} from 'src/lib/interactionManager/types'
 import I18n from 'src/locales/i18n'
 import strings from 'src/locales/strings'
 import { IconToggle } from 'react-native-material-ui'
